LayoutTests/imported/w3c/web-platform-tests/eventsource/format-mime-valid-bogus.any-expected.txt

CONSOLE MESSAGE: EventSource's response has a MIME type ("text/x-bogus") that is not "text/event-stream". Aborting the connection.

PASS EventSource: incorrect valid MIME type
